requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
python-dotenv==0.9.1
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.14.1
